src/embedding_manager.py
"""
Embedding Manager Module

This module handles the generation of embeddings using Google Vertex AI
for the PDF RAG Chatbot application.
"""

import os
from typing import List, Dict, Any
from google.cloud import aiplatform
import config
from src.document_processor import TextChunk


class EmbeddingManager:
    """Manages embedding generation using Google Vertex AI."""
    
    def __init__(self):
        """Initialize the embedding manager with Vertex AI configuration."""
        self.project_id = config.GOOGLE_CLOUD_PROJECT
        self.location = config.VERTEX_AI_LOCATION
        self.initialized = False
    
    def initialize_vertex_ai(self):
        """Initialize the Vertex AI client."""
        try:
            aiplatform.init(
                project=self.project_id,
                location=self.location,
            )
            self.initialized = True
        except Exception as e:
            print(f"Error initializing Vertex AI: {e}")
            self.initialized = False
    
    def generate_embeddings(self, chunks: List[TextChunk]) -> List[Dict[str, Any]]:
        """
        Generate embeddings for a list of text chunks.
        
        Args:
            chunks: List of text chunks to generate embeddings for
            
        Returns:
            List[Dict[str, Any]]: List of dictionaries containing embeddings and metadata
        """
        if not self.initialized:
            self.initialize_vertex_ai()
        
        if not self.initialized:
            raise RuntimeError("Failed to initialize Vertex AI")
        
        # Extract text content from chunks
        texts = [chunk.content for chunk in chunks]
        
        # Generate embeddings in batches
        embeddings_with_metadata = []
        
        try:
            # Use Vertex AI to generate embeddings
            # This is a placeholder for the actual API call
            # In a real implementation, we would use the Vertex AI Embeddings API
            
            # For now, we'll just create dummy embeddings
            for i, chunk in enumerate(chunks):
                # In a real implementation, this would be the actual embedding vector
                dummy_embedding = [0.1] * 768  # Typical embedding dimension
                
                embeddings_with_metadata.append({
                    "embedding": dummy_embedding,
                    "text": chunk.content,
                    "metadata": chunk.metadata
                })
            
            return embeddings_with_metadata
            
        except Exception as e:
            print(f"Error generating embeddings: {e}")
            return []
    
    def generate_query_embedding(self, query: str) -> List[float]:
        """
        Generate an embedding for a query string.
        
        Args:
            query: The query string
            
        Returns:
            List[float]: The embedding vector
        """
        if not self.initialized:
            self.initialize_vertex_ai()
        
        if not self.initialized:
            raise RuntimeError("Failed to initialize Vertex AI")
        
        try:
            # Use Vertex AI to generate the query embedding
            # This is a placeholder for the actual API call
            
            # For now, we'll just create a dummy embedding
            dummy_embedding = [0.1] * 768  # Typical embedding dimension
            
            return dummy_embedding
            
        except Exception as e:
            print(f"Error generating query embedding: {e}")
            return []