requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-google-vertexai>=0.1.0
google-cloud-aiplatform>=1.38.0
PyPDF2>=3.0.1
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0